% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{cv_tree}
\alias{cv_tree}
\title{Title}
\usage{
cv_tree(tree, chars_test, root, nsim = 10, cl = 1)
}
\arguments{
\item{tree}{an object of the class "phylo" which the function will evaluate}

\item{chars_test}{a matrix in which each row name is
a tip of "tree" and each column is a hold-out character (typically not used in the construction of tree)}

\item{root}{the tip of "tree" that should be used as a root}

\item{nsim}{number of Monte Carlo simulations to be used by simmap. Defaults to 10}

\item{cl}{Number of cores to be used. Defaults to 1}
}
\value{
A list with two components: mean_score, the estimated predictive score (smaller is better),
 and sd_score, its standard error
}
\description{
Title
}
\examples{

set.seed(0)
my_tree<- ape::rtree(n=26,tip.label=LETTERS)
chars_test <- replicate(2,ape::rTraitDisc(my_tree,rate = 0.5))
cv_tree(my_tree,chars_test,root="A")
}
